application_factory/requirements.txt
# Core Framework Dependencies
streamlit>=1.28.0
langgraph>=0.0.40
langchain>=0.1.0
langchain-google-genai>=1.0.0

# RAG and Vector Store Dependencies
faiss-cpu>=1.7.4
langchain-community>=0.0.20

# PDF Generation
reportlab>=4.0.0

# Utility Dependencies
python-dotenv>=1.0.0
typing-extensions>=4.8.0

# Text Processing
tiktoken>=0.5.0

# Development Dependencies (optional)
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0 